Split rest package into multiple files
