--- server/adapters/rest/api.go
@@ -1,18 +1,0 @@
-package rest
-
-// const urlParamOrgId string = "orgId"
-
-// func NewServer(c Config) (Server, error) {
-
-// 	s.router.Route("/api", func(r chi.Router) {
-
-// 		r.Group(func(r chi.Router) {
-// 				r.Post("/", s.handleCreateNewOrg())
-// 				r.Route(fmt.Sprintf("/{%s}", urlParamOrgId), func(r chi.Router) {
-// 				})
-// 			})
-// 		})
-// 	})
-
-// 	return s, nil
-// }
